+++ src/lib/pages/module-details/components/tables/history/PublishedEventsTable.tsx
@@ -3,7 +3,7 @@ import { TableContainer } from "@chakra-ui/react";
 import { useMobile } from "lib/app-provider";
 import { Loading } from "lib/components/Loading";
 import { MobileTableContainer } from "lib/components/table";
-import type { ModuleHistory } from "lib/services/move/module";
+import type { ModuleHistory } from "lib/services/types";
 import type { Option } from "lib/types";
 
 import { PublishedEventsTableHeader } from "./PublishedEventsTableHeader";